backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0
pydantic[email]==2.5.0
python-multipart==0.0.6 